+++ application/src/main/java/org/opentripplanner/standalone/config/routerequest/RouteRequestConfig.java
@@ -841,7 +841,9 @@ public class RouteRequestConfig {
         c
           .of("stairsReluctance")
           .since(V2_0)
-          .summary("Used instead of walkReluctance for stairs.")
+          .summary(
+            "A multiplier to specify how bad walking on stairs is, on top of the reluctance parameter."
+          )
           .asDouble(dft.stairsReluctance())
       )
       .withStairsTimeFactor(