+++ doc/user/Changelog.md
@@ -14,6 +14,7 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Show GBFS Geofencing zones in debug UI [#6876](https://github.com/opentripplanner/OpenTripPlanner/pull/6876)
 - Add OTP Badge to ServerInfo [#6865](https://github.com/opentripplanner/OpenTripPlanner/pull/6865)
 - Apply correct traversal permissions to indoor tagged entities [#6923](https://github.com/opentripplanner/OpenTripPlanner/pull/6923)
+- Multiplicative stairs and turn reluctance [#6783](https://github.com/opentripplanner/OpenTripPlanner/pull/6783)
 [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)
 
 ## 2.8.1 (2025-09-22)